scripts use ${BASH_SOURCE[0]##*/} instead of $THIS_FILE_NAME
Each script at its vary beginning was setting its own filename into
variable THIS_FILE_NAME but sourcing others scripts that value was
changed messing up the result inside the main if/then/fi. It was not a
problem but the temptation to use such variable also inside was a source
of possible and nasty mistakes. So, a bashism has been used instead.

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>